backend/requirements.txt
# AI Data Analyst - Python Dependencies

# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# AI/ML
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
